+++ libs/core/date-picker/date-picker.model.ts
@@ -7,6 +7,6 @@ export interface DatePicker<D> extends MobileMode {
     selectedDate: Nullable<D>;
     isOpenChange: EventEmitter<boolean>;
     dialogApprove(): void;
-    dialogDismiss(value: D | DateRange<D>): void;
-    getSelectedDate(): D | DateRange<D>;
+    dialogDismiss(value: D | Array<D> | DateRange<D> | Array<DateRange<D>>): void;
+    getSelectedDate(): D | Array<D> | DateRange<D> | Array<DateRange<D>>;
 }